almost done with homepage, added javascript animations
