+++ app/assets/stylesheets/slides.js
@@ -23,9 +23,15 @@ slideArea.addEventListener("click", function () {
 
 slideArea.addEventListener("mouseover", function () {
   images.forEach(image => {
-    const x = 100 * Math.random() - 50
-    const y = 100 * Math.random() - 50
+    const x = 25 * (Math.floor(Math.random() * 5)) - 50
+    const y = 25 * (Math.floor(Math.random() * 5)) - 50
 
-    image.style.transform = 'translate($(x)px, $(y)px)'
+    image.style.transform = `translate(${x}px, ${y}px)`
+  })
+})
+
+slideArea.addEventListener("mouseout", function (){
+  images.forEach(image => {
+    image.style.transform = ""
   })
 })
