+++ app/javascript/packs/application.js
@@ -25,10 +25,12 @@ require("channels")
 // External imports
 import "bootstrap";
 
+
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
 
+import { initSlides } from '../plugins/slides.js';
+
 document.addEventListener('turbolinks:load', () => {
-  // Call your functions here, e.g:
-  // initSelect2();
-});
+  initSlides();
+})
