+++ app/views/offers/show.html.erb
@@ -8,4 +8,3 @@
 <div class="artist-card">
 <!-- image place holder -->
 <button> <%#= button_to "Book", { action: "booking confirmation path placeholder", class: "btn btn-primary"}, :method => :get %> </button>
-
